Hand of Cards, Great OOP Base
I think this was a great tutorial on oop and I feel like it will be a great file to continue to test my knowledge on.  Hopefully, I'll be expanding into Blackjack, Texas hold'em, and Go Fish.  This will also be a good way for me to test ideas about running programs that use multiple files.